:lipstick:[Design] CommentComponent 생성 및 분리 #246
