+++ open-market/src/pages/product/ProductDetail.tsx
@@ -1,7 +1,13 @@
+import CommentListItem, {
+	CommentBlock,
+	CommentContainer,
+	CommentInputForm,
+	CommentTextarea,
+	ShowStarRating,
+} from "@/components/CommentComponent";
 import { loggedInState } from "@/states/authState";
-import { Common } from "@/styles/common";
+
 import { axiosInstance, debounce } from "@/utils";
-import styled from "@emotion/styled";
 import AccountCircleIcon from "@mui/icons-material/AccountCircle";
 import BookmarkOutlinedIcon from "@mui/icons-material/BookmarkOutlined";
 import CheckIcon from "@mui/icons-material/Check";
@@ -16,142 +22,6 @@ import { Helmet } from "react-helmet-async";
 import toast from "react-hot-toast";
 import { Link, useNavigate, useParams } from "react-router-dom";
 import { useRecoilValue } from "recoil";
-
-interface CommentProps {
-	user?: boolean;
-}
-
-const CommentContainer = styled.article`
-	width: 1160px;
-	min-height: 346px;
-	padding: ${Common.space.spacingMd};
-	margin: ${Common.space.spacingXl} auto;
-	display: flex;
-	flex-flow: column nowrap;
-	justify-content: space-between;
-	position: relative;
-	border: 1px solid ${Common.colors.gray};
-	border-radius: 10px;
-
-	h3 {
-		position: absolute;
-		top: -40px;
-		left: 0;
-
-		& :first-of-type {
-			margin-right: 5px;
-			position: relative;
-			top: 6px;
-			color: ${Common.colors.gray2};
-		}
-	}
-
-	.moreButton {
-		width: 100px;
-		height: 40px;
-		position: relative;
-		left: 50%;
-		transform: translateX(-50%);
-		background-color: transparent;
-		border: none;
-		font-weight: ${Common.font.weight.regular};
-
-		&::after {
-			content: "";
-			position: absolute;
-			top: 50%;
-			transform: translateY(-30%);
-			right: 12px;
-			border-bottom: solid 8px transparent;
-			border-top: solid 8px black;
-			border-left: solid 8px transparent;
-			border-right: solid 8px transparent;
-		}
-	}
-`;
-
-const CommentListItem = styled.li`
-	width: 100%;
-	margin: ${Common.space.spacingMd} 0;
-	display: flex;
-	flex-flow: column nowrap;
-	position: relative;
-`;
-
-const CommentBlock = styled.div<CommentProps>`
-	width: ${(props) => (props.user ? "260px" : "100%")};
-	min-height: ${(props) => (props.user ? "24px" : "40px")};
-	height: auto;
-	white-space: pre-wrap;
-	word-break: normal;
-	overflow-wrap: break-word;
-	padding: ${(props) =>
-		props.user ? `0 ${Common.space.spacingMd}` : `${Common.space.spacingMd}`};
-	display: flex;
-	flex-flow: column nowrap;
-	border: 1px solid ${Common.colors.gray};
-	border-radius: 10px;
-	${(props) =>
-		!props.user &&
-		`
-    line-height: ${Common.space.spacingXl};
-  `};
-
-	${(props) =>
-		props.user &&
-		`
-    justify-content: center;
-    position: absolute;
-    top: 1px;
-    left: ${Common.space.spacingXl};
-  `};
-
-	& :last-of-type {
-		display: ${(props) => (props.user ? "relative" : "flex")};
-		justify-content: ${(props) => (props.user ? "" : "flex-end")};
-	}
-`;
-
-const CommentInputForm = styled.form`
-	position: relative;
-
-	.a11yHidden {
-		display: ${Common.a11yHidden};
-	}
-
-	.commentTextAreaContainer {
-		width: 100%;
-		height: 80px;
-		border: 1px solid ${Common.colors.gray};
-		border-radius: 10px;
-	}
-
-	.inputRating {
-		position: absolute;
-		top: -2px;
-		left: 300px;
-	}
-
-	button {
-		width: 100px;
-		height: 24px;
-		position: absolute;
-		bottom: 5px;
-		right: 5px;
-		color: ${Common.colors.black};
-		font-weight: ${Common.font.weight.bold};
-		background-color: ${Common.colors.emphasize};
-		border: none;
-		border-radius: 10px;
-	}
-`;
-
-const CommentTextarea = styled(CommentBlock)`
-	height: 50px;
-	line-height: normal;
-	word-break: normal;
-	border-width: 0;
-`.withComponent("textarea");
 
 function ProductDetail() {
 	const navigate = useNavigate();
@@ -267,17 +137,6 @@ function ProductDetail() {
 		if (confirm("로그인 후 이용 가능합니다")) {
 			navigate("/signin");
 		}
-	}
-
-	function ShowStarRating({ rating }: { rating: number }) {
-		return (
-			<Rating
-				name="showRating"
-				value={Number(rating)}
-				precision={0.5}
-				readOnly
-			/>
-		);
 	}
 
 	useEffect(() => {
@@ -500,24 +359,12 @@ function ProductDetail() {
 						</CommentInputForm>
 					)}
 				</div>
-
 				<ul>
 					{product?.replies?.length === 0 ? (
 						<p>댓글이 없습니다.</p>
 					) : (
 						product?.replies?.map((reply) => {
-							return (
-								<CommentListItem key={reply._id}>
-									<div>
-										<AccountCircleIcon />
-										<CommentBlock user>{reply.userName}</CommentBlock>
-									</div>
-									<CommentBlock>
-										<span>{reply.content}</span>
-										<ShowStarRating rating={reply.rating} />
-									</CommentBlock>
-								</CommentListItem>
-							);
+							return <CommentListItem reply={reply} />;
 						})
 					)}
 				</ul>
